+++ libs/ui/examples/webpack.js/CHANGELOG.md
@@ -1,11 +1,18 @@
 # webpack.js-example
 
-## 0.2.6-nightly.0
+## 0.2.6
 
 ### Patch Changes
 
 - Updated dependencies []:
-  - @ledgerhq/react-ui@0.35.1-nightly.0
+  - @ledgerhq/react-ui@0.35.1
+
+## 0.2.6-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/react-ui@0.35.1-next.0
 
 ## 0.2.5