+++ libs/ui/packages/react/CHANGELOG.md
@@ -1,11 +1,18 @@
 # @ledgerhq/react-ui
 
-## 0.35.1-nightly.0
+## 0.35.1
 
 ### Patch Changes
 
 - Updated dependencies [[`96654b0`](https://github.com/LedgerHQ/ledger-live/commit/96654b01611519901cb1957213c154f8bcda599c)]:
-  - @ledgerhq/crypto-icons-ui@1.17.0-nightly.0
+  - @ledgerhq/crypto-icons-ui@1.17.0
+
+## 0.35.1-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`96654b0`](https://github.com/LedgerHQ/ledger-live/commit/96654b01611519901cb1957213c154f8bcda599c)]:
+  - @ledgerhq/crypto-icons-ui@1.17.0-next.0
 
 ## 0.35.0